stop_motor.sh: added if condition to check if process is found

+++ scripts/stop_motor.sh
@@ -1,5 +1,13 @@
 #!/bin/bash
 
-kill -9 `pgrep -x controller`
-kill -9 `pgrep -x motor_monitor`
-kill -9 `pgrep -x motor`
+if [ ! -z `pgrep -x controller` ]; then
+    kill -9 `pgrep -x controller`
+fi
+
+if [ ! -z `pgrep -x motor_monitor` ]; then
+    kill -9 `pgrep -x motor_monitor`
+fi
+
+if [ ! -z `pgrep -x motor` ]; then
+    kill -9 `pgrep -x motor`
+fi
